Use json output instead plain-text
Openstack CLI utilities can return warning messages while
run to stderr. It's a normal behavior for Openstack CLI.
But Puppet mixs stderr and stdout for commands.

For preventing parsing warning messages,
I propose got response from Neutron CLI as JSON if it required
and parse it by ruby JSON parser.

JSON is a more usable format for interprocess communication,
than plain-text or CSV.

Change-Id: Iba1105bc119272715fa211ac129a366fe50cbf47
Closes-bug: #1508511

[Fuel9-Kilo] NOTE: this is backport from upstream (liberty+) puppet-neutron module.
It is needed not only but for proper TLS support